Refactor load_data and analyze_data. Add to daily aggregate graph.
